requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pypdf
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community